Detect update available on first stream connection.
Give each app bundle a unique bundle_id.

Sets server_id to process.env.SERVER_ID if set, or else the bundle_id.

Delivers the server_id to the client in the app html (via Meteor
runtime config), so stream_client can detect that an update is
available on the first connection.