Low-level route policy.
Low level package for declaring which routes should serve app html and
which are network.  (This package does not implement server-side
routes or a router, but a higher level router package would call this
package to declare that certain routes should now serve app html).

Modifies stream and accounts-oauth-helper to declare /sockjs/ and
/_oauth/ as network routes.

The server now only serves app html for routes declared to serve the
app; and returns 404 by default for routes that no middleware is
serving.

Used by the appcache package to generate the app cache manifest.